requirements.txt
alembic==1.16.2
blinker==1.9.0
cffi==1.17.1
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.5
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.3.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
